modify collection and project models
